♻️ Refactor[#6]: 사용하는 폴더의 .gitkeep 제거 및 각 파일 rename
