+++ src/main/java/com/saeparam/HeyRoutine/domain/user/repository/UserRepository.java
@@ -1,5 +1,5 @@
 package com.saeparam.HeyRoutine.user.repository;
-import com.saeparam.HeyRoutine.user.domain.User;
+import com.saeparam.HeyRoutine.domain.user.entity.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.Optional;
