+++ src/main/java/com/saeparam/HeyRoutine/global/security/jwt/JwtAuthenticationFilter.java
@@ -1,8 +1,8 @@
 package com.saeparam.HeyRoutine.user.filter;
 
 
-import com.saeparam.HeyRoutine.exception.handler.TokenHandler;
-import com.saeparam.HeyRoutine.response.ApiResponse;
+import com.saeparam.HeyRoutine.global.error.handler.TokenHandler;
+import com.saeparam.HeyRoutine.global.web.response.ApiResponse;
 import com.saeparam.HeyRoutine.user.token.JwtTokenProvider;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import jakarta.servlet.FilterChain;
